services/notification-service/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
boto3==1.33.0
marshmallow==3.20.1
